# backend/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: PickIT print request platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterBody"
      responses:
        "201":
          description: Registration successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Invalid input or email already taken
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /orders:
    post:
      operationId: createOrder
      tags: [orders]
      summary: Create a new print order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderBody"
      responses:
        "201":
          description: Order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /orders/{id}/status:
    patch:
      operationId: updateOrderStatus
      tags: [orders]
      summary: Update order status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOrderStatusBody"
      responses:
        "200":
          description: Updated order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    RegisterBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
        role:
          type: string
          enum: [student, owner]
        shopName:
          type: string
        shopAddress:
          type: string
      required:
        - name
        - email
        - password
        - role
    LoginBody:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required:
        - email
        - password
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
      required:
        - token
        - user
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [student, owner]
        shopId:
          type: integer
          nullable: true
      required:
        - id
        - name
        - email
        - role
...
    CreateOrderBody:
      type: object
      properties:
        shopId:
          type: integer
        fileUrl:
          type: string
        fileName:
          type: string
        pages:
          type: integer
        colorMode:
          type: string
          enum: [bw, color]
        copies:
          type: integer
        note:
          type: string
      required:
        - shopId
        - fileUrl
        - fileName
        - pages
        - colorMode
        - copies
    Order:
      type: object
      properties:
        id:
          type: integer
        studentId:
          type: integer
        shopId:
          type: integer
        fileUrl:
          type: string
        fileName:
          type: string
        pages:
          type: integer
        colorMode:
          type: string
          enum: [bw, color]
        copies:
          type: integer
        status:
          type: string
          enum: [pending, printing, ready, done]
        price:
          type: number
        note:
          type: string
          nullable: true
        createdAt:
          type: string
        studentName:
          type: string
          nullable: true
      required:
        - id
        - studentId
        - shopId
        - fileUrl
        - fileName
        - pages
        - colorMode
        - copies
        - status
        - price
        - createdAt
    UpdateOrderStatusBody:
      type: object
      properties:
        status:
          type: string
          enum: [pending, printing, ready, done]
      required:
        - status
...